pyproject.toml
[tool.poetry]
name = "env-data-retrieval"
version = "0.1.0"
description = ""
authors = ["Qingchuan Ma <qingchuan.ma@nuance.com>"]

[tool.poetry.dependencies]
python = "^3.7"
openpyxl = "^2.6"
python-docx = "^0.8.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
